+++ config/initializers/devise.rb
@@ -255,6 +255,7 @@ Devise.setup do |config|
                       idp_cert_fingerprint: active_provider.providable.idp_cert_fingerprint,
                       certificate: active_provider.providable.sp_certificate,
                       private_key: active_provider.providable.sp_private_key,
+                      uid_attribute: (active_provider.providable.respond_to?(:uid_attribute) && active_provider.providable.uid_attribute.present?) ? active_provider.providable.uid_attribute : nil,
                       security: OneLogin::RubySaml::Settings::DEFAULTS[:security].merge({
                         authn_requests_signed: active_provider.providable.authn_requests_signed,
                         want_assertions_signed: active_provider.providable.want_assertions_signed