+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+- improvement: add uid_attribute for saml provider
+
 ## v6.3.25 2024 May 28
 
 - Fix a bug: member wallet amount calculation error in the excel export